sample pulling apart LDAP roles
