+++ ui/app/adapters/ldap/role/dynamic.js
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: BUSL-1.1
+ */
+
+import LdapRoleAdapter from '../role';
+
+export default class LdapDynamicAdapter extends LdapRoleAdapter {}
