+++ ui/app/serializers/ldap/role/static.js
@@ -1,0 +1,8 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: BUSL-1.1
+ */
+
+import LdapRoleSerializer from '../role';
+
+export default class LdapRoleStaticSerializer extends LdapRoleSerializer {}
